Track norm, linear, and attn layer scratch space
